+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.7.0] - 2023-01-15
+
+### Added
+- support configurable graphql paths
+- support multiple queries (Apollo)
+
+### Changed
+- full scanning the graphql query to parse the operation name.
+
+### Fixed
+- ignore whitespaces at start of operation name
+
 ## [3.6.0] - 2022-11-17
 
 ### Added